==> analysis/scripts/3_3_game_learning.ipynb <==
from utils import *

# Minigame round differences

data = pd.read_csv(os.path.join('..', 'data', 'minigame_results.csv'), delimiter=';')
data = data[data.respondent_id != 1].reset_index(drop=True)
data.note = data.note.map(lambda x: x[x.find('_')+1:])
data.head()

data['order'] = -1
for window in data.rolling(15):
    if(len(window) == 15 and window.index.values[0] % 15 == 0):
        data.loc[window.index.values, 'order'] = [0, 0, 0, 1, 1, 1, 2, 2, 2, 3, 3, 3, 4, 4, 4]

my_shapiro(data, ['hits', 'points'], verbose=True)

## Points and hits differences

data[['hits', 'points', 'game_number']].groupby(['game_number']).agg(
    MeanHits=('hits', np.mean),
    StdHits=('hits', np.std),
    MeanPoints=('points', np.mean),
    StdPoints=('points', np.std),
)

my_friedman(data, ['hits', 'points'], 'game_number', [1, 2, 3])

my_kruskal(data, ['hits', 'points'], 'game_number', [1, 2, 3])

### Ordered rounds

my_friedman(data[data.order == 0], ['hits', 'points'], 'game_number', [1, 2, 3])

data[data.order == 0][['hits', 'points', 'game_number']].groupby(['game_number']).agg(
    MeanHits=('hits', np.mean),
    StdHits=('hits', np.std),
    MeanPoints=('points', np.mean),
    StdPoints=('points', np.std),
)

my_friedman(data[data.order == 4], ['hits', 'points'], 'game_number', [1, 2, 3])

data[data.order == 4][['hits', 'points', 'game_number']].groupby(['game_number']).agg(
    MeanHits=('hits', np.mean),
    StdHits=('hits', np.std),
    MeanPoints=('points', np.mean),
    StdPoints=('points', np.std),
)

my_friedman(data[data.order == 2], ['hits', 'points'], 'game_number', [1, 2, 3])

### Individual mice

configs = {'micr1000_sys10' : [0,4], 'dell2400_sys5' : [1,3], 'dell2400_sys10':None, 'dell2400_sys18' : [1,3],'trust4800_sys10': [0,4]}

for config in configs:
    print(config)
    print(my_kruskal(data[data.note == config], ['hits', 'points'], 'game_number', [1, 2, 3]), end='\n\n')

for config in configs:
    print(config)
    print(data[(data.note==config)][['hits', 'points', 'game_number']].groupby(['game_number']).agg(
        MeanHits=('hits', np.mean),
        StdHits=('hits', np.std),
        MeanPoints=('points', np.mean),
        StdPoints=('points', np.std),
    ), end='\n\n')

for order in [0, 1, 2, 3, 4]:
    for config in configs:
        res = my_kruskal(data[(data.order == order) & (data.note == config)], ['hits', 'points'], 'game_number', [1, 2, 3])
        if(res.loc[0, 'p_value'] != 'nan'):
            print('order:', order, '  config:', config)
            print(res, end='\n\n')

### Results across trials (out of 5) for each mouse

for config in configs:
    if(configs[config]):
        print(config)
        print()
        print(my_mannwhitney(data[(data.note==config)], ['hits', 'points'], 'order', configs[config]), end='\n\n')

for config in configs:
    if(configs[config]):
        for round in [1, 2, 3]:
            print('config: ', config, '   round: ', round)
            print()
            print(my_mannwhitney(data[(data.note==config) & (data.game_number==round)], ['hits', 'points'], 'order', configs[config]), end='\n\n')

for config in configs:
    print(config)
    print(data[(data.note==config)][['hits', 'points', 'order', 'game_number']].groupby(['order', 'game_number']).agg(
        MeanHits=('hits', np.mean),
        StdHits=('hits', np.std),
        MeanPoints=('points', np.mean),
        StdPoints=('points', np.std),
    ), end='\n\n')

### Points - anova

my_shapiro(data[data.game_number == 1], ['points'], True)
my_shapiro(data[data.game_number == 2], ['points'], True)
my_shapiro(data[data.game_number == 3], ['points'], True)

print(AnovaRM(data=data[['session_id', 'points', 'game_number']].reset_index(drop=True).reset_index(drop=True), depvar='points', subject='session_id', within=['game_number']).fit())
print('1vs2:', 2.25031e-12)
print('1vs3:', 1.33227e-15)
print('2vs3:', 0.00638247)  

## Errors

data_raw = pd.read_csv(os.path.join('..', 'data', 'mouseLogs_minigame.csv'), delimiter=';')

data_raw = data_raw[(data_raw['event'] != 'mousemove')]
data_raw = data_raw.sort_values(by=['respondent_id', 'session_id', 'game_number', 'timestamp'])
data_grouped = data_raw.groupby(['respondent_id', 'session_id', 'game_number'])

hits_and_errors = pd.DataFrame(columns=['respondent_id', 'session_id', 'game_number', 'note', 'sex', 'calc_hits', 'errors'])

for name, group in data_grouped :
    hits = 1
    errors = 0
    next_target_number = group.head(1)['next_target_number'].values[0]

    for i, item in group.iterrows():
        if (item['event'] == 'mouseup'):
            if (item['next_target_number'] == next_target_number + 1):
                hits += 1
                next_target_number += 1
            elif(item['next_target_number'] == next_target_number):
                errors += 1
    row = [name[0], name[1], name[2], group['note'].max(), group['sex'].max(), hits, errors]
    hits_and_errors.loc[len(hits_and_errors)] = row

my_friedman(hits_and_errors, ['errors'], 'game_number', [1, 2, 3])
